chore: create a seperate component for inter protocol route's fallback ui
